Add files via upload
Here's my Digital Image Processing project on "Multiresolution Bilateral Filtering for Image Denoising". The bilateral filter is used for picture-denoising, it is a nonlinear filter that performs spatial averaging without smoothing edges. An important issue with the bilateral filter is the selection of filter parameters, which affects the results significantly. The contribution of this paper is a multiresolution bilateral filter, which extends the concept of bilateral filtering by applying bilateral filtering to the approximate (low-frequency) subbands of a signal that has been wavelet filter bank decomposed. A novel picture-denoising framework is created by combining the multiresolution bilateral filter with wavelet thresholding, and it proves to be highly successful in removing noise from real noisy images.